requirements.txt
absl-py==0.15.0
anyio==3.5.0
appdirs==1.4.4
astunparse==1.6.3
audioread==2.1.9
cached-property==1.5.2
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==7.1.2
cycler==0.11.0
decorator==5.1.1
dill==0.3.4
ffmpeg==1.4
ffmpeg-python==0.2.0
flatbuffers==1.12
fonttools==4.29.1
future==0.18.2
gast==0.4.0
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h11==0.12.0
h2==4.1.0
h5py==3.1.0
hpack==4.0.0
httpcore==0.13.7
httpx==0.19.0
hyperframe==6.0.1
idna==3.3
importlib-metadata==3.10.1
joblib==1.1.0
keras==2.8.0
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
librosa==0.8.0
llvmlite==0.36.0
Markdown==3.3.4
matplotlib==3.5.1
norbert==0.2.1
numba==0.53.1
numpy==1.21.5
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
Pillow==9.0.1
pooch==1.6.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.2.2
rfc3986==1.5.0
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
six==1.15.0
sklearn==0.0
sniffio==1.2.0
SoundFile==0.10.3.post1
spleeter==2.3.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-estimator==2.5.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.8
Werkzeug==2.0.3
wrapt==1.12.1
zipp==3.7.0
